Update README.md
Gave a much needed lift to the README for the master debacle change.
Remember the hundred of thousands of whites who died in Americas little civil war. Where's the rewards for those dencendants? And the appropiate triggers and all that?

Here's a little fact of reality: You alone is responsible for your actions and the future, relying on people who are dead for clout is fucking pathethic. No excuses. If you want to join civilization, act civilized. It's that simple. If I spent all day making excuses I'd get nothing done, exactly like you.